pyproject.toml
[tool.poetry]
name = "integration-api"
version = "0.1.0"
description = "API adapter for product integrations to a marketplace platform"
authors = ["Kilian Zimmerer"]

[tool.poetry.dependencies]
python = "^3.8"
mangum = "^0.15.0"
fastapi = "^0.78.0"
uvicorn = "^0.18.1"
pytest = "^7.1.2"
flake8 = "^4.0.1"
black = "^22.3.0"
python-dotenv = "^0.20.0"
requests = "^2.28.0"
python-multipart = "^0.0.5"
pandas = "^1.4.3"
pytest-vcr = "^1.0.2"
aiohttp = "^3.8.1"

[tool.poetry.dev-dependencies]
pytest-watch = "^4.2.0"
pytest-asyncio = "^0.18.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.pytest.ini_options]
asyncio_mode = 'strict'
